requirements.txt
jinja2==3.1.6
markdown==3.8.2
pygments==2.19.2
python-frontmatter==1.1.0
pyyaml==6.0.2
